Add minimum time flag for processlist metrics
Add a flag for controlling the minimum time that a thread must be in a
specific state in order to be counted.

This allows for detection of threads "stuck" in a specific state for a
minimum period of time.  This avoids counting short-lived thread state
chagnes from polluting the metrics.