Fix a mixture of error messages and metrics is returned from the multiple target endpoint.
If the config file doesn't contain a section requested by the
`auth_module` parameter and/or a network address requested by the
`target` parameter is ill-formed, emit one line of error message only.

This changes not-yet-released behaviour that emits both error messages
and metrics.

This commit also removes `probe_success` metrics, also not yet released,
for the following reasons:

  - Existing `/metrics` endpoint does not contain it.
  - It will always be 1 even if failed to connect to the
    MySQL server.
  - Existing `mysql_up` parameter can be used instead.

Signed-off-by: Kohei YOSHIDA <kohei@yosida95.com>